video/GUI - Fullstack/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
opencv-python-headless
numpy
Pillow
